basic "circle power" fix (float imperfection => double)

+++ core/src/com/darkgran/smc/play/CircleActor.java
@@ -8,9 +8,9 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 public class CircleActor extends Actor {
     private final LevelStage levelStage;
     private CircleBody circleBody;
-    private float radius;
+    private double radius;
 
-    public CircleActor(final LevelStage levelStage, float x, float y, float radius, BodyDef.BodyType bodyType) {
+    public CircleActor(final LevelStage levelStage, float x, float y, double radius, BodyDef.BodyType bodyType) {
         this.levelStage = levelStage;
         circleBody = new CircleBody(levelStage.getWorldScreen().getWorld(), this, x, y, radius, bodyType);
         this.radius = radius;
@@ -20,16 +20,16 @@ public class CircleActor extends Actor {
     public void drawShape(ShapeRenderer shapeRenderer, Color color) {
         shapeRenderer.begin(ShapeRenderer.ShapeType.Filled);
         shapeRenderer.setColor(color);
-        int segments = Math.round(getRadius()*200);
+        int segments = Math.round((float) getRadius()*200);
         if (segments < 10) { segments = 10; }
         else if (segments > 100) { segments = 50; }
-        shapeRenderer.circle(getCircleBody().getBody().getPosition().x, getCircleBody().getBody().getPosition().y, getRadius(), segments);
+        shapeRenderer.circle(getCircleBody().getBody().getPosition().x, getCircleBody().getBody().getPosition().y, (float) getRadius(), segments);
         shapeRenderer.setColor(Color.WHITE);
         shapeRenderer.end();
     }
 
     public void refreshActorBounds() {
-        this.setBounds(circleBody.getBody().getPosition().x-radius, circleBody.getBody().getPosition().y-radius, radius*2, radius*2);
+        this.setBounds(circleBody.getBody().getPosition().x-(float) radius, circleBody.getBody().getPosition().y-(float) radius, (float) radius*2, (float) radius*2);
     }
 
     public CircleBody getCircleBody() {
@@ -40,11 +40,11 @@ public class CircleActor extends Actor {
         this.circleBody = circleBody;
     }
 
-    public float getRadius() {
+    public double getRadius() {
         return radius;
     }
 
-    public void setRadius(float radius) {
+    public void setRadius(double radius) {
         this.radius = radius;
     }
